+++ reechauto-server-news/src/main/java/com/reechauto/cloud/news/bean/req/privilege/MenuQueryByParentIdRequest.java
@@ -7,14 +7,12 @@ public class MenuQueryByParentIdRequest extends BaseRequest {
 
 	private static final long serialVersionUID = 1L;
 	@NotNull(message = "pId不可以为null")
-	private int pId;
-
-	public int getpId() {
+	private Long pId;
+	public Long getpId() {
 		return pId;
 	}
-
-	public void setpId(int pId) {
+	public void setpId(Long pId) {
 		this.pId = pId;
 	}
-	
+
 }
