+++ reechauto-server-news/src/main/java/com/reechauto/cloud/news/entity/SysMenuExample.java
@@ -114,43 +114,43 @@ public class SysMenuExample {
             addCriterion("id is not null");
             return (Criteria) this;
         }
-        public Criteria andIdEqualTo(Integer value) {
+        public Criteria andIdEqualTo(Long value) {
             addCriterion("id =", value, "id");
             return (Criteria) this;
         }
-        public Criteria andIdNotEqualTo(Integer value) {
+        public Criteria andIdNotEqualTo(Long value) {
             addCriterion("id <>", value, "id");
             return (Criteria) this;
         }
-        public Criteria andIdGreaterThan(Integer value) {
+        public Criteria andIdGreaterThan(Long value) {
             addCriterion("id >", value, "id");
             return (Criteria) this;
         }
-        public Criteria andIdGreaterThanOrEqualTo(Integer value) {
+        public Criteria andIdGreaterThanOrEqualTo(Long value) {
             addCriterion("id >=", value, "id");
             return (Criteria) this;
         }
-        public Criteria andIdLessThan(Integer value) {
+        public Criteria andIdLessThan(Long value) {
             addCriterion("id <", value, "id");
             return (Criteria) this;
         }
-        public Criteria andIdLessThanOrEqualTo(Integer value) {
+        public Criteria andIdLessThanOrEqualTo(Long value) {
             addCriterion("id <=", value, "id");
             return (Criteria) this;
         }
-        public Criteria andIdIn(List<Integer> values) {
+        public Criteria andIdIn(List<Long> values) {
             addCriterion("id in", values, "id");
             return (Criteria) this;
         }
-        public Criteria andIdNotIn(List<Integer> values) {
+        public Criteria andIdNotIn(List<Long> values) {
             addCriterion("id not in", values, "id");
             return (Criteria) this;
         }
-        public Criteria andIdBetween(Integer value1, Integer value2) {
+        public Criteria andIdBetween(Long value1, Long value2) {
             addCriterion("id between", value1, value2, "id");
             return (Criteria) this;
         }
-        public Criteria andIdNotBetween(Integer value1, Integer value2) {
+        public Criteria andIdNotBetween(Long value1, Long value2) {
             addCriterion("id not between", value1, value2, "id");
             return (Criteria) this;
         }
@@ -218,43 +218,43 @@ public class SysMenuExample {
             addCriterion("p_id is not null");
             return (Criteria) this;
         }
-        public Criteria andPIdEqualTo(Integer value) {
+        public Criteria andPIdEqualTo(Long value) {
             addCriterion("p_id =", value, "pId");
             return (Criteria) this;
         }
-        public Criteria andPIdNotEqualTo(Integer value) {
+        public Criteria andPIdNotEqualTo(Long value) {
             addCriterion("p_id <>", value, "pId");
             return (Criteria) this;
         }
-        public Criteria andPIdGreaterThan(Integer value) {
+        public Criteria andPIdGreaterThan(Long value) {
             addCriterion("p_id >", value, "pId");
             return (Criteria) this;
         }
-        public Criteria andPIdGreaterThanOrEqualTo(Integer value) {
+        public Criteria andPIdGreaterThanOrEqualTo(Long value) {
             addCriterion("p_id >=", value, "pId");
             return (Criteria) this;
         }
-        public Criteria andPIdLessThan(Integer value) {
+        public Criteria andPIdLessThan(Long value) {
             addCriterion("p_id <", value, "pId");
             return (Criteria) this;
         }
-        public Criteria andPIdLessThanOrEqualTo(Integer value) {
+        public Criteria andPIdLessThanOrEqualTo(Long value) {
             addCriterion("p_id <=", value, "pId");
             return (Criteria) this;
         }
-        public Criteria andPIdIn(List<Integer> values) {
+        public Criteria andPIdIn(List<Long> values) {
             addCriterion("p_id in", values, "pId");
             return (Criteria) this;
         }
-        public Criteria andPIdNotIn(List<Integer> values) {
+        public Criteria andPIdNotIn(List<Long> values) {
             addCriterion("p_id not in", values, "pId");
             return (Criteria) this;
         }
-        public Criteria andPIdBetween(Integer value1, Integer value2) {
+        public Criteria andPIdBetween(Long value1, Long value2) {
             addCriterion("p_id between", value1, value2, "pId");
             return (Criteria) this;
         }
-        public Criteria andPIdNotBetween(Integer value1, Integer value2) {
+        public Criteria andPIdNotBetween(Long value1, Long value2) {
             addCriterion("p_id not between", value1, value2, "pId");
             return (Criteria) this;
         }
